refactor(124123): :hammer: move mocks para src/
